refactor(server): split xmpp call tests
